refactor(gen-image): remove dead code and dedupe image model handlers
- Delete never-imported matureWords.ts (581 lines) + utils/mature.ts
- Delete promptEnhancer.ts (identity no-op) and strip the dead
  accept-language/enhance + pimpPrompt path from normalizeAndTranslatePrompt;
  drop wasPimped, TimingStep/timingInfo, and MinimalRequest plumbing
- Remove no-op createBar/errorBar stubs from ProgressManager
- Drop unused logError consts in xai/qwen/wan image handlers
- Remove 12 log-and-rethrow catch blocks in the generateImage switch
  (errors already mapped by handler.ts throwImageError); keep the 3
  logGptImageError blocks, drop their dead updateBar error line
- Drop speculative concurrentRequests param plumbing + countFluxJobs/countJobs
  (a per-request KV round-trip that was only logged)
- Extract shared DashScope multimodal helper (qwen + wan image)
- Extract shared xAI aspect-ratio table/helper (image + video)
