+++ gen.pollinations.ai/src/image/availableServers.ts
@@ -93,12 +93,6 @@ export async function getRegisteredServers(
     );
 }
 
-export async function countJobs(type: ServerType = "flux"): Promise<number> {
-    return (await getRegisteredServers(type)).length;
-}
-
-export const countFluxJobs = () => countJobs("flux");
-
 export const getNextServerUrl = async (
     type: ServerType = "flux",
 ): Promise<string> => {
