--- gen.pollinations.ai/src/image/config/matureWords.ts
@@ -1,581 +1,0 @@
-// Note: Some common words removed to reduce false positives
-export const words = [
-    // A
-    "a55",
-    "a_s_s",
-    "anal",
-    "anus",
-    "ar5e",
-    "arrse",
-    "arse",
-    "ass",
-    "ass-fucker",
-    "assfucker",
-    "assfukka",
-    "asshole",
-    "assholes",
-    "asswhole",
-    "ahegao",
-    "adult",
-    "aphrodisiac",
-
-    // B
-    "b!tch",
-    "b00bs",
-    "b17ch",
-    "b1tch",
-    "b4stard",
-    "ballbag",
-    "balls",
-    "ballsack",
-    "bastard",
-    "beastial",
-    "beastiality",
-    "bellend",
-    "bestial",
-    "bestiality",
-    "bi+ch",
-    "biatch",
-    "bimbo",
-    "bitch",
-    "bitcher",
-    "bitchers",
-    "bitches",
-    "bitchin",
-    "bitching",
-    "bloody",
-    "blow job",
-    "blowjob",
-    "blowjobs",
-    "boiolas",
-    "bollock",
-    "bollok",
-    "boner",
-    "boob",
-    "boobs",
-    "booobs",
-    "boooobs",
-    "booooobs",
-    "booooooobs",
-    "bra",
-    "breast",
-    "breasts",
-    "buceta",
-    "bugger",
-    "bum",
-    "bunny fucker",
-    "busty",
-    "butt",
-    "butthole",
-    "buttmuch",
-    "buttplug",
-    "bukake",
-    "bukkake",
-
-    // C
-    "c0ck",
-    "c0cksucker",
-    "camgirl",
-    "carpet muncher",
-    "cawk",
-    "child",
-    "chink",
-    "cipa",
-    "cl1t",
-    "clit",
-    "clitoris",
-    "clits",
-    "cnut",
-    "cock",
-    "cock-sucker",
-    "cockface",
-    "cockhead",
-    "cockmunch",
-    "cockmuncher",
-    "cocks",
-    "cocksuck",
-    "cocksucked",
-    "cocksucker",
-    "cocksucking",
-    "cocksucks",
-    "cocksuka",
-    "cocksukka",
-    "cok",
-    "cokmuncher",
-    "coksucka",
-    "coon",
-    "cox",
-    "crap",
-    "creampie",
-    "cum",
-    "cummer",
-    "cumming",
-    "cums",
-    "cumshot",
-    "cunilingus",
-    "cunillingus",
-    "cunnilingus",
-    "cunt",
-    "cuntlick",
-    "cuntlicker",
-    "cuntlicking",
-    "cunts",
-    "cyalis",
-    "cyberfuc",
-    "cyberfuck",
-    "cyberfucked",
-    "cyberfucker",
-    "cyberfuckers",
-    "cyberfucking",
-    "cuckold",
-    "climax",
-
-    // D
-    "d1ck",
-    "daughter",
-    "deepthroat",
-    "dick",
-    "dickhead",
-    "dildo",
-    "dildos",
-    "dink",
-    "dinks",
-    "dirsa",
-    "dlck",
-    "dog-fucker",
-    "doggin",
-    "dogging",
-    "donkeyribber",
-    "doosh",
-    "duche",
-    "dyke",
-    "dominatrix",
-    "dominate",
-    "deep-throat",
-
-    // E
-    "ejaculate",
-    "ejaculated",
-    "ejaculates",
-    "ejaculating",
-    "ejaculatings",
-    "ejaculation",
-    "ejakulate",
-    "erotic",
-    "explicit",
-    "escort",
-
-    // F
-    "f u c k",
-    "f u c k e r",
-    "f4nny",
-    "fag",
-    "fagging",
-    "faggitt",
-    "faggot",
-    "faggs",
-    "fagot",
-    "fagots",
-    "fags",
-    "fanny",
-    "fannyflaps",
-    "fannyfucker",
-    "fanyy",
-    "fatass",
-    "father",
-    "fcuk",
-    "fcuker",
-    "fcuking",
-    "feck",
-    "fecker",
-    "felching",
-    "fellate",
-    "fellatio",
-    "fingerfuck",
-    "fingerfucked",
-    "fingerfucker",
-    "fingerfuckers",
-    "fingerfucking",
-    "fingerfucks",
-    "fistfuck",
-    "fistfucked",
-    "fistfucker",
-    "fistfuckers",
-    "fistfucking",
-    "fistfuckings",
-    "fistfucks",
-    "flange",
-    "fook",
-    "fooker",
-    "fuck",
-    "fucka",
-    "fucked",
-    "fucker",
-    "fuckers",
-    "fuckhead",
-    "fuckheads",
-    "fuckin",
-    "fucking",
-    "fuckings",
-    "fuckingshitmotherfucker",
-    "fuckme",
-    "fucks",
-    "fuckwhit",
-    "fuckwit",
-    "fudge packer",
-    "fudgepacker",
-    "fuk",
-    "fuker",
-    "fukker",
-    "fukkin",
-    "fuks",
-    "fukwhit",
-    "fukwit",
-    "fux",
-    "fux0r",
-    "f_u_c_k",
-    "facial",
-    "femdom",
-    "fornicate",
-    "footjob",
-
-    // G
-    "gangbang",
-    "gangbanged",
-    "gangbangs",
-    "gaylord",
-    "gaysex",
-    "genital",
-    "girl",
-    "God",
-    "god-dam",
-    "god-damned",
-    "goddamn",
-    "goddamned",
-    "goatse",
-    "grope",
-    "gstring",
-
-    // H
-    "hardcoresex",
-    "heshe",
-    "hitler",
-    "hoar",
-    "hoare",
-    "hoer",
-    "hore",
-    "horny",
-    "horniest",
-    "hotsex",
-    "hustler",
-    "hentai",
-    "handjob",
-    "hump",
-
-    // I
-    "intercourse",
-    "incest",
-    "impregnate",
-
-    // J
-    "jack-off",
-    "jackoff",
-    "jerk-off",
-    "jew",
-    "jism",
-    "jiz",
-    "jizm",
-    "jizz",
-    "juggs",
-    "jerking",
-
-    // K
-    "kawk",
-    "knobhead",
-    "kock",
-    "kum",
-    "kummer",
-    "kumming",
-    "kums",
-    "kunilingus",
-    "kamasutra",
-
-    // L
-    "l3i+ch",
-    "l3itch",
-    "labia",
-    "lesbian",
-    "lingerie",
-    "lmfao",
-    "lust",
-    "lusting",
-    "lapdance",
-    "latex",
-
-    // M
-    "m0f0",
-    "m0fo",
-    "m45terbate",
-    "ma5terb8",
-    "ma5terbate",
-    "masochist",
-    "master-bate",
-    "masterb8",
-    "masterbat*",
-    "masterbat3",
-    "masterbate",
-    "masterbation",
-    "masterbations",
-    "masturbate",
-    "masturbating",
-    "milf",
-    "mo-fo",
-    "mof0",
-    "mofo",
-    "mother fucker",
-    "motherfuck",
-    "motherfucked",
-    "motherfucker",
-    "motherfuckers",
-    "motherfuckin",
-    "motherfucking",
-    "motherfuckings",
-    "motherfuckka",
-    "motherfucks",
-    "mothafuck",
-    "mothafucka",
-    "mothafuckas",
-    "mothafuckaz",
-    "mothafucked",
-    "mothafucker",
-    "mothafuckers",
-    "mothafuckin",
-    "mothafucking",
-    "mothafuckings",
-    "mothafucks",
-    "muff",
-    "mutha",
-    "muthafecker",
-    "muthafuckker",
-    "muther",
-    "mutherfucker",
-    "moaning",
-
-    // N
-    "n1gga",
-    "n1gger",
-    "naked",
-    "nazi",
-    "nigg3r",
-    "nigg4h",
-    "nigga",
-    "niggah",
-    "niggas",
-    "niggaz",
-    "nigger",
-    "niggers",
-    "nob jokey",
-    "nobhead",
-    "nobjocky",
-    "nobjokey",
-    "nsfw",
-    "nude",
-    "numbnuts",
-    "nutsack",
-    "nympho",
-
-    // O
-    "orgasim",
-    "orgasims",
-    "orgasm",
-    "orgasms",
-    "orgy",
-    "onlyfans",
-
-    // P
-    "p0rn",
-    "panties",
-    "pecker",
-    "penis",
-    "penisfucker",
-    "phonesex",
-    "phuck",
-    "phuk",
-    "phuked",
-    "phuking",
-    "phukked",
-    "phukking",
-    "phuks",
-    "phuq",
-    "pigfucker",
-    "pimpis",
-    "piss",
-    "pissed",
-    "pisser",
-    "pissers",
-    "pisses",
-    "pissflaps",
-    "pissin",
-    "pissing",
-    "pissoff",
-    "poop",
-    "porn",
-    "porno",
-    "pornography",
-    "pornos",
-    "pornstars",
-    "pregnant",
-    "prick",
-    "pricks",
-    "pron",
-    "pube",
-    "pusse",
-    "pussi",
-    "pussies",
-    "pussy",
-    "pussys",
-    "playboy",
-    "pegging",
-    "prostitute",
-
-    // Q
-    "quickie",
-
-    // R
-    "rape",
-    "raping",
-    "rectum",
-    "retard",
-    "rimjaw",
-    "rimming",
-    "roughsex",
-
-    // S
-    "s hit",
-    "s.o.b.",
-    "schlong",
-    "screwing",
-    "scroat",
-    "scrote",
-    "scrotum",
-    "semen",
-    "sex",
-    "sexy",
-    "sh!+",
-    "sh!t",
-    "sh1t",
-    "shag",
-    "shagger",
-    "shaggin",
-    "shagging",
-    "shemale",
-    "shi+",
-    "shit",
-    "s_h_i_t",
-    "shitdick",
-    "shite",
-    "shited",
-    "shitey",
-    "shitfuck",
-    "shitfull",
-    "shithead",
-    "shiting",
-    "shitings",
-    "shits",
-    "shitted",
-    "shitter",
-    "shitters",
-    "shitting",
-    "shittings",
-    "shitty",
-    "skank",
-    "slut",
-    "sluts",
-    "smut",
-    "smegma",
-    "snatch",
-    "son",
-    "son-of-a-bitch",
-    "spac",
-    "spanking",
-    "spunk",
-    "sister",
-    "strip",
-    "stripper",
-    "sensual",
-    "suck",
-    "sucking",
-    "suckoff",
-    "suckjob",
-
-    // T
-    "t1tt1e5",
-    "t1tties",
-    "teen",
-    "teenage",
-    "teenager",
-    "teets",
-    "teez",
-    "testical",
-    "testicle",
-    "titfuck",
-    "tits",
-    "titt",
-    "tittie5",
-    "tittiefucker",
-    "titties",
-    "tittyfuck",
-    "tittywank",
-    "titwank",
-    "tosser",
-    "topless",
-    "turd",
-    "tw4t",
-    "twat",
-    "twathead",
-    "twatty",
-    "twunt",
-    "twunter",
-    "threesome",
-    "throatfuck",
-
-    // U
-    "upskirt",
-    "uncensored",
-
-    // V
-    "v14gra",
-    "v1gra",
-    "vagina",
-    "viagra",
-    "vulva",
-    "voyeur",
-
-    // W
-    "w00se",
-    "waifu",
-    "wang",
-    "wank",
-    "wanker",
-    "wanky",
-    "whore",
-    "whoar",
-    "willies",
-    "willy",
-    "woman",
-    "wetdream",
-    "wetpussy",
-
-    // X
-    "xrated",
-    "xxx",
-
-    // Y
-    "years",
-    "year",
-    "young",
-    "yiff",
-
-    // Z
-    "zoophilia",
-];
