+++ gen.pollinations.ai/src/image/models/wanImageModel.ts
@@ -4,15 +4,10 @@ import { getImageEnv } from "../env.ts";
 import { HttpError } from "../httpError.ts";
 import type { ImageParams } from "../params.ts";
 import type { ProgressManager } from "../progressBar.ts";
-import { fetchUpstream } from "../utils/fetchUpstream.ts";
+import { callDashScopeMultimodalImage } from "../utils/dashScopeImage.ts";
 import { downloadUserImage } from "../utils/imageDownload.ts";
 
 const logOps = debug("pollinations:wan-image:ops");
-const logError = debug("pollinations:wan-image:error");
-
-// DashScope multimodal generation endpoint (synchronous)
-const DASHSCOPE_API_BASE =
-    "https://dashscope-intl.aliyuncs.com/api/v1/services/aigc/multimodal-generation/generation";
 
 // Wan 2.7 model IDs on DashScope
 const WAN_IMAGE_MODEL = "wan2.7-image";
@@ -51,29 +46,6 @@ function clampDimensions(
     }
 
     return [w, h];
-}
-
-interface WanImageResponse {
-    output?: {
-        choices?: Array<{
-            finish_reason: string;
-            message: {
-                role: string;
-                content: Array<{ image?: string; type?: string }>;
-            };
-        }>;
-        finished?: boolean;
-    };
-    usage?: {
-        image_count?: number;
-        size?: string;
-        input_tokens?: number;
-        output_tokens?: number;
-        total_tokens?: number;
-    };
-    request_id?: string;
-    code?: string;
-    message?: string;
 }
 
 /**
@@ -161,7 +133,7 @@ export async function callWanImageAPI(
         parameters,
     };
 
-    return callDashScopeWanImageAPI(
+    return callDashScopeMultimodalImage(
         apiKey,
         requestBody,
         isPro ? "wan-image-pro" : "wan-image",
@@ -169,82 +141,4 @@ export async function callWanImageAPI(
         progress,
         requestId,
     );
-}
-
-/**
- * Call the DashScope multimodal generation API and return the result
- */
-async function callDashScopeWanImageAPI(
-    apiKey: string,
-    requestBody: Record<string, unknown>,
-    actualModel: string,
-    modelLabel: string,
-    progress: ProgressManager,
-    requestId: string,
-): Promise<ImageGenerationResult> {
-    // Log request safely (hide base64 data)
-    const safeBody = JSON.stringify(requestBody, (key, value) => {
-        if (
-            key === "image" &&
-            typeof value === "string" &&
-            value.startsWith("data:")
-        ) {
-            return "[base64]";
-        }
-        return value;
-    });
-    logOps("DashScope Wan image request:", safeBody);
-
-    const response = await fetchUpstream(DASHSCOPE_API_BASE, {
-        method: "POST",
-        headers: {
-            Authorization: `Bearer ${apiKey}`,
-            "Content-Type": "application/json",
-        },
-        body: JSON.stringify(requestBody),
-        errorLabel: `${modelLabel} API failed`,
-    });
-
-    const data: WanImageResponse = await response.json();
-
-    if (data.code) {
-        throw new HttpError(
-            `${modelLabel} error: ${data.message || data.code}`,
-            400,
-            undefined,
-            DASHSCOPE_API_BASE,
-        );
-    }
-
-    const imageUrl = data.output?.choices?.[0]?.message?.content?.[0]?.image;
-    if (!imageUrl) {
-        throw new HttpError(
-            `No image URL in ${modelLabel} response`,
-            500,
-            undefined,
-            DASHSCOPE_API_BASE,
-        );
-    }
-
-    logOps("Image generated, downloading from:", imageUrl);
-    progress.updateBar(requestId, 80, "Processing", "Downloading image...");
-
-    const imageResponse = await fetchUpstream(imageUrl, {
-        errorLabel: "Failed to download generated image",
-    });
-    const buffer = Buffer.from(await imageResponse.arrayBuffer());
-    logOps(`Image downloaded, size: ${(buffer.length / 1024).toFixed(1)} KB`);
-    progress.updateBar(requestId, 95, "Success", "Image generation completed");
-
-    return {
-        buffer,
-        isMature: false,
-        isChild: false,
-        trackingData: {
-            actualModel,
-            usage: {
-                completionImageTokens: 1, // flat per-image pricing
-            },
-        },
-    };
 }
