+++ gen.pollinations.ai/src/image/progressBar.ts
@@ -1,8 +1,6 @@
 export class ProgressManager {
-    createBar(_id: string, _title: string) {}
     updateBar(_id: string, _progress: number, _step: string, _status: string) {}
     completeBar(_id: string, _status = "Complete") {}
-    errorBar(_id: string, _error: string) {}
     stop() {}
     startRequest(_id: string) {
         return this;
